+++ packages/symbiote-ui/canvas/Breadcrumb/Breadcrumb.tpl.js
@@ -1,7 +1,0 @@
-/**
- * Breadcrumb template
- * @module symbiote-node/canvas/Breadcrumb.tpl
- */
-import { html } from '@symbiotejs/symbiote';
-
-export let template = html` <div ${{ itemize: 'crumbs', 'item-tag': 'breadcrumb-item' }}></div> `;