+++ packages/symbiote-ui/display/Banner/Banner.tpl.js
@@ -1,3 +1,0 @@
-import { html } from '@symbiotejs/symbiote';
-
-export default html`<slot></slot>`;