+++ packages/symbiote-ui/effects/CellBg/CellBg.tpl.js
@@ -1,5 +1,0 @@
-import { html } from '@symbiotejs/symbiote';
-
-export default html`
-  <canvas ref="canvas"></canvas>
-`;