+++ packages/symbiote-ui/node/NodeSocket/NodeSocket.tpl.js
@@ -1,7 +1,0 @@
-/**
- * NodeSocket template
- * @module symbiote-node/node/NodeSocket.tpl
- */
-import { html } from '@symbiotejs/symbiote';
-
-export let template = html`<slot></slot>`;